// data/github/current/ui/packages/issue-viewer/components/sections/development-section/__generated__/DevelopmentPickerQuery.graphql.ts
export type DevelopmentPickerQuery$variables = {
  id: string;
};
export type DevelopmentPickerQuery$data = {
  readonly node: {
    readonly repository?: {
      readonly name: string;
      readonly nameWithOwner: string;
      readonly owner: {
        readonly login: string;
      };
    };
  } | null | undefined;
};
export type DevelopmentPickerQuery = {
  response: DevelopmentPickerQuery$data;
  variables: DevelopmentPickerQuery$variables;
};